+++ packages/jaspr_cli/lib/src/migrations/build_method_migration.dart
@@ -286,36 +286,28 @@ class BuilderVisitor extends RecursiveAstVisitor<void> {
     if (node.parent case FunctionExpression(
       parent: NamedArgument(
         name: Token(lexeme: 'builder'),
-        parent: ArgumentList(
-          parent: MethodInvocation(
-            methodName: SimpleIdentifier(
-              name: 'Builder' ||
-                  'StatefulBuilder' ||
-                  'AsyncBuilder' ||
-                  'ListenableBuilder' ||
-                  'StreamBuilder' ||
-                  'FutureBuilder',
-            ),
+        parent: ArgumentList(parent: final MethodInvocation invocation),
+      ),
+    )) {
+      switch (invocation) {
+        case MethodInvocation(
+          methodName: SimpleIdentifier(
+            name: 'Builder' ||
+                'StatefulBuilder' ||
+                'AsyncBuilder' ||
+                'ListenableBuilder' ||
+                'StreamBuilder' ||
+                'FutureBuilder',
           ),
-        ),
-      ),
-    )) {
-      onBuilderFunction(node);
-      return;
-    }
-
-    if (node.parent case FunctionExpression(
-      parent: NamedArgument(
-        name: Token(lexeme: 'builder'),
-        parent: ArgumentList(parent: final MethodInvocation m),
-      ),
-    )) {
-      if (m case MethodInvocation(
-        methodName: SimpleIdentifier(name: 'single'),
-        target: SimpleIdentifier(name: 'Builder'),
-      )) {
-        onSingleBuilder(m);
-        return;
+        ):
+          onBuilderFunction(node);
+          return;
+        case MethodInvocation(
+          methodName: SimpleIdentifier(name: 'single'),
+          target: SimpleIdentifier(name: 'Builder'),
+        ):
+          onSingleBuilder(invocation);
+          return;
       }
     }